MCP: better debug traces for button press
git-svn-id: svn://localhost/ardour2/branches/3.0@11903 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/surfaces/mackie/surface.cc
@@ -361,7 +361,7 @@ Surface::handle_midi_pitchbend_message (MIDI::Parser&, MIDI::pitchbend_t pb, uin
 void 
 Surface::handle_midi_note_on_message (MIDI::Parser &, MIDI::EventTwoBytes* ev)
 {
-	DEBUG_TRACE (DEBUG::MackieControl, string_compose ("SurfacePort::handle_note_on %1 = %2\n", ev->note_number, ev->velocity));
+	DEBUG_TRACE (DEBUG::MackieControl, string_compose ("SurfacePort::handle_note_on %1 = %2\n", (int) ev->note_number, (int) ev->velocity));
 	
 	Button* button = buttons[ev->note_number];
